Refactor JdbcUtil into JdbcUri and TypedDataSourceConfiguration
